+++ VoxelEngine/VoxelEngine/VoxelEngine/CoreEngine/Input.cpp
@@ -118,16 +118,16 @@ void Input::mouse_pos_cb([[maybe_unused]] GLFWwindow* window, double xpos, doubl
   static bool firstMouse = true;
   if (firstMouse)
   {
-    screenOffset.x = xpos;
-    screenOffset.y = ypos;
+    screenOffset.x = static_cast<float>(xpos);
+    screenOffset.y = static_cast<float>(ypos);
     firstMouse = false;
   }
 
   screenPos.x = static_cast<float>(xpos);
   screenPos.y = static_cast<float>(ypos);
 
-  screenOffset.x = xpos - prevScreenPos.x;
-  screenOffset.y = prevScreenPos.y - ypos;
+  screenOffset.x = static_cast<float>(xpos) - prevScreenPos.x;
+  screenOffset.y = prevScreenPos.y - static_cast<float>(ypos);
   prevScreenPos = glm::vec2(xpos, ypos);
   screenOffset *= sensitivity;
